requirements.txt
altair==5.5.0
annotated-types @ file:///C:/b/abs_0dmaoyhhj3/croot/annotated-types_1709542968311/work
anyio @ file:///C:/b/abs_31grv7a7bl/croot/anyio_1729121290992/work
attrs==25.1.0
blinker==1.9.0
cachetools==5.5.1
certifi @ file:///C:/b/abs_59o0xj7aav/croot/certifi_1734473304008/work/certifi
charset-normalizer==3.4.1
click==8.1.8
colorama @ file:///C:/Users/dev-admin/perseverance-python-buildout/croot/colorama_1699472650914/work
distro @ file:///C:/b/abs_71xr36ua5r/croot/distro_1714488282676/work
gitdb==4.0.12
GitPython==3.1.44
h11 @ file:///C:/b/abs_1czwoyexjf/croot/h11_1706652332846/work
httpcore @ file:///C:/b/abs_55n7g233bw/croot/httpcore_1706728507241/work
httpx @ file:///C:/b/abs_43e135shby/croot/httpx_1723474830126/work
idna @ file:///C:/b/abs_aad84bnnw5/croot/idna_1714398896795/work
Jinja2==3.1.5
jiter @ file:///C:/b/abs_acsxrw36a6/croot/jiter_1729808947806/work
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
markdown-it-py==3.0.0
MarkupSafe==3.0.2
mdurl==0.1.2
narwhals==1.24.2
numpy==2.2.2
openai @ file:///C:/b/abs_fbl6noof_d/croot/openai_1738233380947/work
packaging==24.2
pandas==2.2.3
pillow==11.1.0
protobuf==5.29.3
pyarrow==19.0.0
pydantic @ file:///C:/b/abs_27dx58x550/croot/pydantic_1734736090499/work
pydantic_core @ file:///C:/b/abs_bdosz7qwys/croot/pydantic-core_1734726071532/work
pydeck==0.9.1
Pygments==2.19.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2025.1
referencing==0.36.2
requests==2.32.3
rich==13.9.4
rpds-py==0.22.3
setuptools==75.1.0
six==1.17.0
smmap==5.0.2
sniffio @ file:///C:/b/abs_3akdewudo_/croot/sniffio_1705431337396/work
streamlit==1.41.1
tenacity==9.0.0
toml==0.10.2
tornado==6.4.2
tqdm @ file:///C:/b/abs_77wju137gk/croot/tqdm_1724853945787/work
typing_extensions @ file:///C:/b/abs_0ffjxtihug/croot/typing_extensions_1734714875646/work
tzdata==2025.1
urllib3==2.3.0
watchdog==6.0.0
wheel==0.44.0
